requirements.txt
pymongo
requests
pandas
sentence_transformers
python-dotenv
einops
openai